Improve queue loading
also separate multi_ops policy.
