+++ spec/spec_helper.rb
@@ -19,6 +19,6 @@ RSpec.configure do |config|
   config.expect_with(:rspec) { |c| c.syntax = :expect }
   config.order = "random"
 
-  config.before(:all) {}
-  config.after {}
+  config.before { Postqueue.reset! }
+  config.after  { }
 end
